feat: 현재 내가 로그인 상태인지 확인하는 함수 구현
인증 서비스에 isLoined라는 함수를 추가함
나의 상태를 알고 페이지 렌더링을 하기 위해
